Extract interface for service layer
