+++ bin/bompare.dart
@@ -8,7 +8,8 @@ import 'package:bompare/persistence/parser/wss_inventory_result_parser.dart';
 import 'package:bompare/persistence/report_writer.dart';
 import 'package:bompare/persistence/scan_result_loader.dart';
 import 'package:bompare/service/bom_service.dart';
-import 'package:bompare/service/domain/business_exception.dart';
+import 'package:bompare/service/business_exception.dart';
+import 'package:bompare/service/domain/bom_interactor.dart';
 
 void main(List<String> arguments) async {
   final loader = ScanResultLoader(
@@ -19,7 +20,7 @@ void main(List<String> arguments) async {
     },
   );
   final reporter = ReportWriter();
-  final service = BomService(loader, reporter);
+  final service = BomInteractor(loader, reporter);
 
   try {
     final command = CommandRunner('bompare', 'Bill-Of-Material scan comparator')
